+++ ninaAPI/WebService/V2/Common.cs
@@ -1,0 +1,8 @@
+namespace ninaAPI.WebService.V2.Equipment
+{
+    public interface INinaWatcher
+    {
+        public void StartWatchers();
+        public void StopWatchers();
+    }
+}